[feat] change Global func UserImage to Class UserImageService
